Fix XSS potential with server rendering
Adding a vote (or title) with `</script>` will end the script tag in the
`<head>` for `__INITIAL_STATE__`.  Everything after that is just HTML,
so it's possible to add a further `<script>` tag to load / set external
JavaScript running in the context of the page.

To see the issue, add a choice with the vote text
`Hello world </script><script>alert('Hi there')</script>`

Then go to the page that displays the vote options. Refresh the page,
and observe the 'Hi there' alert box.

This change uses the safer-stringify npm package to escape the
`</script>`.