Fix prerender issue for new votes
When a new vote is created, the Choice gets a `percent` property, which
the first time results in `null` (or possibly `NaN`), because it does a
`100 / totalVotes`, which is a divide by zero.

This ends up with the vote getting a `width:null%` or `width:NaN%`
depending on server or client, which results in the prerender checksum
not matching.

This commit fixes the percent, so initially, it will be 0%, and removes
the extra `percent` prop from the ChoiceBar in VotingComponent, as it is
then overridden by the `choice.percent` property (due to the
`{...choice}`).